+++ apis/core-api/src/core/controllers/CityController/CityController.ts
@@ -1,6 +1,6 @@
 import { OrderEnum } from '@ultras/utils';
 import BaseController from 'core/controllers/BaseController';
-import { CityService, CountryService } from 'services';
+import { CityService, CountryService } from 'core/services';
 
 import { DEFAULT_PAGINATION_ATTRIBUTES } from '@constants';
 import { DbIdentifier } from 'types';
