+++ apis/core-api/src/core/controllers/TeamController/TeamController.ts
@@ -1,6 +1,6 @@
 import { OrderEnum } from '@ultras/utils';
 import BaseController from 'core/controllers/BaseController';
-import { CountryService, TeamService } from 'services';
+import { CountryService, TeamService } from 'core/services';
 
 import { DEFAULT_PAGINATION_ATTRIBUTES } from '@constants';
 import { DbIdentifier } from 'types';
