+++ apis/core-api/src/core/services/BaseService.ts
