+++ apis/core-api/src/core/services/CityService.ts
@@ -5,11 +5,11 @@ import {
   DbIdentifier,
 } from 'types';
 
-import BaseService from 'services/BaseService';
 import resources from 'core/data/lcp';
 import db from 'core/data/models';
 import { CityAttributes, CityCreationAttributes } from 'core/data/models/City';
 import injectCities, { RapidApiCity } from 'core/data/inject-scripts/injectCities';
+import BaseService from './BaseService';
 
 export interface CitiesListParamsInterface {
   name?: string;
