+++ apis/core-api/src/core/services/FavoriteTeamService.ts
@@ -1,8 +1,8 @@
 import { DbIdentifier } from 'types';
-import BaseService from 'services/BaseService';
-
 import db from 'core/data/models';
 import { FavoriteTeamCreationAttributes } from 'core/data/models/FavoriteTeam';
+
+import BaseService from './BaseService';
 
 class FavoriteTeamService extends BaseService {
   static async addToUserFavorites(
