+++ apis/core-api/src/core/services/MatchService.ts
@@ -5,12 +5,12 @@ import {
   ServiceByIdResultType,
   DbIdentifier,
 } from 'types';
-
-import BaseService from 'services/BaseService';
 import resources from 'core/data/lcp';
 import db from 'core/data/models';
 import { MatchAttributes, MatchCreationAttributes } from 'core/data/models/Match';
 import injectMatches, { RapidApiMatch } from 'core/data/inject-scripts/injectMatches';
+
+import BaseService from './BaseService';
 
 export interface MatchesListParamsInterface {
   dateFrom?: string;
